use consistent notion for both languages
