+++ vital/CMakeLists.txt
@@ -132,7 +132,7 @@ set( vital_public_headers
   types/track_descriptor_set.h
   types/vector.h
   types/uid.h
-  types/file_name.h
+  types/filename.h
 
 )
 
@@ -184,7 +184,7 @@ set( vital_sources
   types/track_descriptor.cxx
   types/track_set.cxx
   types/uid.cxx
-  types/file_name.cxx
+  types/filename.cxx
 )
 
 kwiver_install_headers(
